// src/modules/partner-portal/ducks/info/types.ts
export enum ActionTypes {
    FETCH_NEWS_REQUEST = 'PartnerPortalModule/info/FETCH_NEWS_REQUEST',
    FETCH_NEWS_SUCCESS = 'PartnerPortalModule/info/FETCH_NEWS_SUCCESS',
    FETCH_NEWS_FAILURE = 'PartnerPortalModule/info/FETCH_NEWS_FAILURE',

    FETCH_STATUS_REQUEST = 'PartnerPortalModule/info/FETCH_STATUS_REQUEST',
    FETCH_STATUS_SUCCESS = 'PartnerPortalModule/info/FETCH_STATUS_SUCCESS',
    FETCH_STATUS_FAILURE = 'PartnerPortalModule/info/FETCH_STATUS_FAILURE',

    SET_NEWS = 'PartnerPortalModule/info/SET_NEWS',
    SET_STATUS = 'PartnerPortalModule/info/SET_STATUS',
}

export interface IState {}